unify venvs with uv + deduplicate code
- Single pyproject.toml + .venv via uv (was 2 separate venvs)
- Merged backend/rules/ → root rules/ (mature codebase with to_dict, thread safety)
- Merged database.py (WAL, indexes, foreign keys, leaderboards + clear_database)
- Moved backend/main.py → web_main.py
- Updated all TUI imports to use rules.xxx
- Removed 6 duplicate root-level files + entire backend/ directory
- Updated dev.sh to use uv run
- Added AI thinking overlay with dark blur + white/red bouncing dots
- Updated README.md and ARCHITECTURE.md

+++ models.py
@@ -30,7 +30,7 @@ class ShoeState(BaseModel):
     remaining: int
     total: int
     running_count: int
-    true_count: int
+    true_count: float
     penetration_pct: float
     played_ranks: dict[str, int] = {}
